[#8] Move error to type
